Added LED to SensorsInterface, changed light to float
Added functions to the SensorsInterface.c/py files to turn on and off
the orange LED on the Sensorian Shield. Changed the ambient light
function to return a float instead of int and convert from a c_type
accordingly. Added a sleep to the end of setupSensorian to ensure the
sensors are ready before proceeding.