Only change the FXOS files for Scratch and NodeRED
Previously changed all the FXOS C drivers in the repository but now see
that there are two different versions of pretty much all the C drivers
that need to be reconciled before I can change them too. Only changed
the one version of the driver which is found in the Scratch and NodeRED
folders.